Fixed vim doing strange hard tabs
For some reason my vim settings got reset and it decided to start doing
hard tabs despite the fact that it was showing me that they were spaces.
I went through and fixed all of the cases I was aware of. This has also
effected commits, so that will be changing hopefully.

  Just a quick commit test right here.